% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ods_api_endpoint.R
\name{set_ods_api_endpoint}
\alias{set_ods_api_endpoint}
\title{Change the default ODS API endpoint}
\usage{
set_ods_api_endpoint(endpoint_url)
}
\arguments{
\item{endpoint_url}{The new URL of the API endpoint. E.g. "https://directory.spineservices.nhs.uk/ORD/2-0-0/"}
}
\value{

}
\description{
The ODS functions in this package will send API queries to
*https://directory.spineservices.nhs.uk/ORD/2-0-0/*. This function can change
the URL if required. Note: Compatability with newer versions of the API
cannot be garanteed.
}
\examples{
\dontrun{
set_ods_api_endpoint("https://directory.spineservices.nhs.uk/ORD/2-0-1/")
}
}
