% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_mortality.R
\docType{data}
\name{ons_mortality}
\alias{ons_mortality}
\title{Deaths registered weekly in England and Wales, provisional}
\format{
Data frame with five columns
\describe{
\item{Category}{Data is released in groups: Totals, by age groups, males by age, females by age,
All respiratory diseases (ICD-10 J00-J99)  ICD-10 v 2001 and by Region}
\item{Counts}{Counts of deaths in whole numbers}
\item{Date}{Provisional deaths are released on the Friday of every week.}
\item{WeekNo}{Following the format of the ONS spreadsheets each week is numbered sequentially within the year.}
}
}
\source{
Collected by Zoë Turner \email{zoe.turner2@nottshc.nhs.uk}, Apr-2020 from \url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/datasets/weeklyprovisionalfiguresondeathsregisteredinenglandandwales}
}
\usage{
data(ons_mortality)
}
\description{
Provisional counts of the number of deaths registered in England and Wales, by age, sex and region, in the latest weeks for which data are available.
}
\details{
Source and licence acknowledgement
This data has been made available through Office of National Statistics under the Open Government
Licence \url{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}
}
\examples{
data(ons_mortality)

library(dplyr)
library(tidyr)

wideForm <- ons_mortality \%>\%
 select(-week_no) \%>\%
 pivot_wider(names_from = date,
             values_from = counts
 )

}
\keyword{England}
\keyword{Provisional}
\keyword{Wales}
\keyword{datasets}
\keyword{deaths}
\keyword{mortality}
