% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_uk_population_2023.R
\docType{data}
\name{ons_uk_population_2023}
\alias{ons_uk_population_2023}
\title{ONS Mid-2023 Population Estimate for UK}
\format{
Tibble with six columns
\describe{
\item{sex}{male or female}
\item{Code}{The country/geography code}
\href{Name}{country of the UK}
\item{Geography}{Country}
\item{age}{year of age}
\item{count}{the number of people in this group}
}
}
\source{
\href{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatesforukenglandandwalesscotlandandnorthernireland}{ONS Estimates of the population for the UK, England, Wales, Scotland, and Northern Ireland}
}
\usage{
data(ons_uk_population_2023)
}
\description{
ONS Population Estimates for Mid-year 2023
National and subnational mid-year population estimates for the UK and its
 constituent countries by administrative area, age and sex (including
  components of population change, median age and population density).
}
\details{
Data sourced from
\href{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatesforukenglandandwalesscotlandandnorthernireland}
}
\examples{
data(ons_uk_population_2023)

library(readxl)
library(tidyverse)
library(tidyr)
Load the data in
population_data_2023_f <- read_excel(
"mye23tablesuk.xlsx",
 sheet="MYE2 - Females",
 skip = 7)
population_data_2023_m <- read_excel(
 "mye23tablesuk.xlsx",
 sheet="MYE2 - Males",
 skip = 7)
pivot longer
population_data_2023_f <- population_data_2023_f |>
 select(!`All ages`) |>
 pivot_longer(`0`:`90+`, names_to = "age", values_to = "count")
population_data_2023_m <- population_data_2023_m |>
 select(!`All ages`) |>
 pivot_longer(`0`:`90+`, names_to = "age", values_to = "count")
population_data_combined <- bind_rows(
 females = population_data_2023_f,
 males = population_data_2023_m,
 .id = "sex"
)



}
\keyword{datasets}
\keyword{ons}
\keyword{population}
